+++ Tests/Wire/Resources/Shipping/OrderDrafts/OrderDraftsClientWireTests.swift
@@ -90,6 +90,28 @@ import Chrt
         try #require(response == expectedResponse)
     }
 
+    @Test func imageToTextV11() async throws -> Void {
+        let stub = HTTPStub()
+        stub.setResponse(
+            body: Data(
+                """
+                string
+                """.utf8
+            )
+        )
+        let client = ChrtClient(
+            baseURL: "https://api.fern.com",
+            token: "<token>",
+            urlSession: stub.urlSession
+        )
+        let expectedResponse = "string"
+        let response = try await client.shipping.orderDrafts.imageToTextV1(
+            request: .init(file: .init(data: Data("".utf8))),
+            requestOptions: RequestOptions(additionalHeaders: stub.headers)
+        )
+        try #require(response == expectedResponse)
+    }
+
     @Test func validateV11() async throws -> Void {
         let stub = HTTPStub()
         stub.setResponse(